fix lambda tests
